+++ .cache/06/023f5482781f66b5bd6b9e6e631f08.json
@@ -1,1 +1,1 @@
-{"id":"img/hello-instagram.png","dependencies":[{"name":"C:\\KN\\LP_DEAB\\package.json","includedInParent":true,"mtime":1560221260569},{"name":"C:\\KN\\LP_DEAB\\.browserslistrc","includedInParent":true,"mtime":1559628543032}],"generated":{"js":"module.exports = \"/f4c525b4b1c42423852b433436871e12.png\";"},"sourceMaps":null,"error":null,"hash":"f62184d96e550935fa3383b84535a777","cacheData":{"env":{}}}
+{"id":"img/hello-instagram.png","dependencies":[{"name":"C:\\KN\\LP_DEAB\\package.json","includedInParent":true,"mtime":1560387452553},{"name":"C:\\KN\\LP_DEAB\\.browserslistrc","includedInParent":true,"mtime":1560387421364}],"generated":{"js":"module.exports = \"/f4c525b4b1c42423852b433436871e12.png\";"},"sourceMaps":null,"error":null,"hash":"f62184d96e550935fa3383b84535a777","cacheData":{"env":{}}}
