+++ .cache/0b/57e538c0a65c9531d3ce0de25e5292.json
@@ -1,1 +1,1 @@
-{"id":"img/works-mountains.png","dependencies":[{"name":"C:\\KN\\LP_DEAB\\package.json","includedInParent":true,"mtime":1560221260569},{"name":"C:\\KN\\LP_DEAB\\.browserslistrc","includedInParent":true,"mtime":1559628543032}],"generated":{"js":"module.exports = \"/f684bfd7fcc5f4976088bc2c7377b155.png\";"},"sourceMaps":null,"error":null,"hash":"f8dc8de0328a45c395675741133447b1","cacheData":{"env":{}}}
+{"id":"img/works-mountains.png","dependencies":[{"name":"C:\\KN\\LP_DEAB\\package.json","includedInParent":true,"mtime":1560387452553},{"name":"C:\\KN\\LP_DEAB\\.browserslistrc","includedInParent":true,"mtime":1560387421364}],"generated":{"js":"module.exports = \"/f684bfd7fcc5f4976088bc2c7377b155.png\";"},"sourceMaps":null,"error":null,"hash":"f8dc8de0328a45c395675741133447b1","cacheData":{"env":{}}}
