+++ .cache/17/55cf145f4b4f39612ce80b7e3ff041.json
@@ -1,1 +1,1 @@
-{"id":"img/works-doors.png","dependencies":[{"name":"C:\\KN\\LP_DEAB\\package.json","includedInParent":true,"mtime":1560221260569},{"name":"C:\\KN\\LP_DEAB\\.browserslistrc","includedInParent":true,"mtime":1559628543032}],"generated":{"js":"module.exports = \"/1f0a5a464ff48bed21a86434dab49420.png\";"},"sourceMaps":null,"error":null,"hash":"dc35cde2f86c2842172b25d5e4261125","cacheData":{"env":{}}}
+{"id":"img/works-doors.png","dependencies":[{"name":"C:\\KN\\LP_DEAB\\package.json","includedInParent":true,"mtime":1560387452553},{"name":"C:\\KN\\LP_DEAB\\.browserslistrc","includedInParent":true,"mtime":1560387421364}],"generated":{"js":"module.exports = \"/1f0a5a464ff48bed21a86434dab49420.png\";"},"sourceMaps":null,"error":null,"hash":"dc35cde2f86c2842172b25d5e4261125","cacheData":{"env":{}}}
