+++ .cache/31/1a998b8f0fd45f8e39ec877858a957.json
@@ -1,1 +1,1 @@
-{"id":"img/gad-google.png","dependencies":[{"name":"C:\\KN\\LP_DEAB\\package.json","includedInParent":true,"mtime":1560221260569},{"name":"C:\\KN\\LP_DEAB\\.browserslistrc","includedInParent":true,"mtime":1559628543032}],"generated":{"js":"module.exports = \"/f0ceea73d3f2046ba33ab8988d567e6f.png\";"},"sourceMaps":null,"error":null,"hash":"f37cbf8dbb4d075795da6ac0b73b7070","cacheData":{"env":{}}}
+{"id":"img/gad-google.png","dependencies":[{"name":"C:\\KN\\LP_DEAB\\package.json","includedInParent":true,"mtime":1560387452553},{"name":"C:\\KN\\LP_DEAB\\.browserslistrc","includedInParent":true,"mtime":1560387421364}],"generated":{"js":"module.exports = \"/f0ceea73d3f2046ba33ab8988d567e6f.png\";"},"sourceMaps":null,"error":null,"hash":"f37cbf8dbb4d075795da6ac0b73b7070","cacheData":{"env":{}}}
