+++ .cache/3d/ecc17d02078adeb242b9a72d1ee347.json
@@ -1,1 +1,1 @@
-{"id":"img/hello-twitter.png","dependencies":[{"name":"C:\\KN\\LP_DEAB\\package.json","includedInParent":true,"mtime":1560221260569},{"name":"C:\\KN\\LP_DEAB\\.browserslistrc","includedInParent":true,"mtime":1559628543032}],"generated":{"js":"module.exports = \"/eeb97d5d125bb3ba4ceb517ff2820e3e.png\";"},"sourceMaps":null,"error":null,"hash":"b0ca180de4fc491361e0374867470286","cacheData":{"env":{}}}
+{"id":"img/hello-twitter.png","dependencies":[{"name":"C:\\KN\\LP_DEAB\\package.json","includedInParent":true,"mtime":1560387452553},{"name":"C:\\KN\\LP_DEAB\\.browserslistrc","includedInParent":true,"mtime":1560387421364}],"generated":{"js":"module.exports = \"/eeb97d5d125bb3ba4ceb517ff2820e3e.png\";"},"sourceMaps":null,"error":null,"hash":"b0ca180de4fc491361e0374867470286","cacheData":{"env":{}}}
