+++ .cache/6f/916d4f6e4d72910017bc1baffa1132.json
@@ -1,1 +1,1 @@
-{"id":"img/works-doors-mobile.png","dependencies":[{"name":"C:\\KN\\LP_DEAB\\package.json","includedInParent":true,"mtime":1560221260569},{"name":"C:\\KN\\LP_DEAB\\.browserslistrc","includedInParent":true,"mtime":1559628543032}],"generated":{"js":"module.exports = \"/a39abf0834dd1dbd80ca49c08de3eb41.png\";"},"sourceMaps":null,"error":null,"hash":"e51a59cc6581444eb6044d0efbec912b","cacheData":{"env":{}}}
+{"id":"img/works-doors-mobile.png","dependencies":[{"name":"C:\\KN\\LP_DEAB\\package.json","includedInParent":true,"mtime":1560387452553},{"name":"C:\\KN\\LP_DEAB\\.browserslistrc","includedInParent":true,"mtime":1560387421364}],"generated":{"js":"module.exports = \"/a39abf0834dd1dbd80ca49c08de3eb41.png\";"},"sourceMaps":null,"error":null,"hash":"e51a59cc6581444eb6044d0efbec912b","cacheData":{"env":{}}}
