+++ .cache/7f/ac0521fb33e70b62249be0f5f2727d.json
@@ -1,1 +1,1 @@
-{"id":"img/header-avatar.png","dependencies":[{"name":"C:\\KN\\LP_DEAB\\package.json","includedInParent":true,"mtime":1560221260569},{"name":"C:\\KN\\LP_DEAB\\.browserslistrc","includedInParent":true,"mtime":1559628543032}],"generated":{"js":"module.exports = \"/9b72d61f84aea2233a6c2df42625bfbc.png\";"},"sourceMaps":null,"error":null,"hash":"18bdc2d518e6ab97e35a100fee34eb70","cacheData":{"env":{}}}
+{"id":"img/header-avatar.png","dependencies":[{"name":"C:\\KN\\LP_DEAB\\package.json","includedInParent":true,"mtime":1560387452553},{"name":"C:\\KN\\LP_DEAB\\.browserslistrc","includedInParent":true,"mtime":1560387421364}],"generated":{"js":"module.exports = \"/9b72d61f84aea2233a6c2df42625bfbc.png\";"},"sourceMaps":null,"error":null,"hash":"18bdc2d518e6ab97e35a100fee34eb70","cacheData":{"env":{}}}
