+++ .cache/8a/3f08629a420f767a6372f3759ed1df.json
@@ -1,1 +1,1 @@
-{"id":"img/gad-dropbox.png","dependencies":[{"name":"C:\\KN\\LP_DEAB\\package.json","includedInParent":true,"mtime":1560221260569},{"name":"C:\\KN\\LP_DEAB\\.browserslistrc","includedInParent":true,"mtime":1559628543032}],"generated":{"js":"module.exports = \"/d00cf2d849076c2e4ffc2ccc08d96d1c.png\";"},"sourceMaps":null,"error":null,"hash":"b1f95c72e49b8abb4277db8e9c3cd89e","cacheData":{"env":{}}}
+{"id":"img/gad-dropbox.png","dependencies":[{"name":"C:\\KN\\LP_DEAB\\package.json","includedInParent":true,"mtime":1560387452553},{"name":"C:\\KN\\LP_DEAB\\.browserslistrc","includedInParent":true,"mtime":1560387421364}],"generated":{"js":"module.exports = \"/d00cf2d849076c2e4ffc2ccc08d96d1c.png\";"},"sourceMaps":null,"error":null,"hash":"b1f95c72e49b8abb4277db8e9c3cd89e","cacheData":{"env":{}}}
