+++ .cache/93/54589c29b685c8b8171bebb5e922cc.json
@@ -1,1 +1,1 @@
-{"id":"img/form-button.png","dependencies":[{"name":"C:\\KN\\LP_DEAB\\package.json","includedInParent":true,"mtime":1560221260569},{"name":"C:\\KN\\LP_DEAB\\.browserslistrc","includedInParent":true,"mtime":1559628543032}],"generated":{"js":"module.exports = \"/135afb0f6bc0851b38101121f7cbb6ca.png\";"},"sourceMaps":null,"error":null,"hash":"400f2c9a4f7a91323a3c5ecb21eebd09","cacheData":{"env":{}}}
+{"id":"img/form-button.png","dependencies":[{"name":"C:\\KN\\LP_DEAB\\package.json","includedInParent":true,"mtime":1560387452553},{"name":"C:\\KN\\LP_DEAB\\.browserslistrc","includedInParent":true,"mtime":1560387421364}],"generated":{"js":"module.exports = \"/135afb0f6bc0851b38101121f7cbb6ca.png\";"},"sourceMaps":null,"error":null,"hash":"400f2c9a4f7a91323a3c5ecb21eebd09","cacheData":{"env":{}}}
