+++ .cache/98/aa0bd3a7f3fc6e709302615dfc65a4.json
@@ -1,1 +1,1 @@
-{"id":"img/works-ocean.png","dependencies":[{"name":"C:\\KN\\LP_DEAB\\package.json","includedInParent":true,"mtime":1560221260569},{"name":"C:\\KN\\LP_DEAB\\.browserslistrc","includedInParent":true,"mtime":1559628543032}],"generated":{"js":"module.exports = \"/bd101e1861e344e3c54f546b025bac14.png\";"},"sourceMaps":null,"error":null,"hash":"4b1f348dead3c9a61da915bc99c00cd9","cacheData":{"env":{}}}
+{"id":"img/works-ocean.png","dependencies":[{"name":"C:\\KN\\LP_DEAB\\package.json","includedInParent":true,"mtime":1560387452553},{"name":"C:\\KN\\LP_DEAB\\.browserslistrc","includedInParent":true,"mtime":1560387421364}],"generated":{"js":"module.exports = \"/bd101e1861e344e3c54f546b025bac14.png\";"},"sourceMaps":null,"error":null,"hash":"4b1f348dead3c9a61da915bc99c00cd9","cacheData":{"env":{}}}
