+++ .cache/a1/f35046333ed50b61f368e49fa38de5.json
@@ -1,1 +1,1 @@
-{"id":"img/hello-behance.png","dependencies":[{"name":"C:\\KN\\LP_DEAB\\package.json","includedInParent":true,"mtime":1560221260569},{"name":"C:\\KN\\LP_DEAB\\.browserslistrc","includedInParent":true,"mtime":1559628543032}],"generated":{"js":"module.exports = \"/f1eb274c65ffdc632ded405cddc063c3.png\";"},"sourceMaps":null,"error":null,"hash":"9aa88951d820b3cf280ad0b3d7712db3","cacheData":{"env":{}}}
+{"id":"img/hello-behance.png","dependencies":[{"name":"C:\\KN\\LP_DEAB\\package.json","includedInParent":true,"mtime":1560387452553},{"name":"C:\\KN\\LP_DEAB\\.browserslistrc","includedInParent":true,"mtime":1560387421364}],"generated":{"js":"module.exports = \"/f1eb274c65ffdc632ded405cddc063c3.png\";"},"sourceMaps":null,"error":null,"hash":"9aa88951d820b3cf280ad0b3d7712db3","cacheData":{"env":{}}}
