+++ .cache/bc/75a90aa3b7ec800764453cdd89db73.json
@@ -1,1 +1,1 @@
-{"id":"img/works-building-mobile.png","dependencies":[{"name":"C:\\KN\\LP_DEAB\\package.json","includedInParent":true,"mtime":1560221260569},{"name":"C:\\KN\\LP_DEAB\\.browserslistrc","includedInParent":true,"mtime":1559628543032}],"generated":{"js":"module.exports = \"/3ff086660fd794b3ca0c8bc6d169e817.png\";"},"sourceMaps":null,"error":null,"hash":"2f9b7d880cb3c07d6ae88c4058f78b21","cacheData":{"env":{}}}
+{"id":"img/works-building-mobile.png","dependencies":[{"name":"C:\\KN\\LP_DEAB\\package.json","includedInParent":true,"mtime":1560387452553},{"name":"C:\\KN\\LP_DEAB\\.browserslistrc","includedInParent":true,"mtime":1560387421364}],"generated":{"js":"module.exports = \"/3ff086660fd794b3ca0c8bc6d169e817.png\";"},"sourceMaps":null,"error":null,"hash":"2f9b7d880cb3c07d6ae88c4058f78b21","cacheData":{"env":{}}}
