+++ .cache/d3/e83da85adeef5cf9ba3b26d75b430b.json
@@ -1,1 +1,1 @@
-{"id":"img/gad-google-mobile.png","dependencies":[{"name":"C:\\KN\\LP_DEAB\\package.json","includedInParent":true,"mtime":1560221260569},{"name":"C:\\KN\\LP_DEAB\\.browserslistrc","includedInParent":true,"mtime":1559628543032}],"generated":{"js":"module.exports = \"/61d18a2da5bd5b13ed440dce7c952006.png\";"},"sourceMaps":null,"error":null,"hash":"c1ceec497dc10ecf5152052f9ad5e8c0","cacheData":{"env":{}}}
+{"id":"img/gad-google-mobile.png","dependencies":[{"name":"C:\\KN\\LP_DEAB\\package.json","includedInParent":true,"mtime":1560387452553},{"name":"C:\\KN\\LP_DEAB\\.browserslistrc","includedInParent":true,"mtime":1560387421364}],"generated":{"js":"module.exports = \"/61d18a2da5bd5b13ed440dce7c952006.png\";"},"sourceMaps":null,"error":null,"hash":"c1ceec497dc10ecf5152052f9ad5e8c0","cacheData":{"env":{}}}
