+++ .cache/e5/8934a2c468f125e2e415ab97c59d4c.json
@@ -1,1 +1,1 @@
-{"id":"img/gad-dropbox-mobile.png","dependencies":[{"name":"C:\\KN\\LP_DEAB\\package.json","includedInParent":true,"mtime":1560221260569},{"name":"C:\\KN\\LP_DEAB\\.browserslistrc","includedInParent":true,"mtime":1559628543032}],"generated":{"js":"module.exports = \"/48936f7f88c51009f8071e0c779c8e45.png\";"},"sourceMaps":null,"error":null,"hash":"bd996b80f9c68c27d43bdac2d5f24be8","cacheData":{"env":{}}}
+{"id":"img/gad-dropbox-mobile.png","dependencies":[{"name":"C:\\KN\\LP_DEAB\\package.json","includedInParent":true,"mtime":1560387452553},{"name":"C:\\KN\\LP_DEAB\\.browserslistrc","includedInParent":true,"mtime":1560387421364}],"generated":{"js":"module.exports = \"/48936f7f88c51009f8071e0c779c8e45.png\";"},"sourceMaps":null,"error":null,"hash":"bd996b80f9c68c27d43bdac2d5f24be8","cacheData":{"env":{}}}
