+++ .cache/eb/3a8a690f266aacaa25e6460ebf7d82.json
@@ -1,1 +1,1 @@
-{"id":"img/works-building.png","dependencies":[{"name":"C:\\KN\\LP_DEAB\\package.json","includedInParent":true,"mtime":1560221260569},{"name":"C:\\KN\\LP_DEAB\\.browserslistrc","includedInParent":true,"mtime":1559628543032}],"generated":{"js":"module.exports = \"/0b6bb3fd543415d6501e202ef8bf7c98.png\";"},"sourceMaps":null,"error":null,"hash":"ac55cdf9f8b20f88143918a456325a05","cacheData":{"env":{}}}
+{"id":"img/works-building.png","dependencies":[{"name":"C:\\KN\\LP_DEAB\\package.json","includedInParent":true,"mtime":1560387452553},{"name":"C:\\KN\\LP_DEAB\\.browserslistrc","includedInParent":true,"mtime":1560387421364}],"generated":{"js":"module.exports = \"/0b6bb3fd543415d6501e202ef8bf7c98.png\";"},"sourceMaps":null,"error":null,"hash":"ac55cdf9f8b20f88143918a456325a05","cacheData":{"env":{}}}
