+++ .cache/ec/3283a09e8b9319e6dd6c420621094b.json
@@ -1,1 +1,1 @@
-{"id":"img/works-mountains-mobile.png","dependencies":[{"name":"C:\\KN\\LP_DEAB\\package.json","includedInParent":true,"mtime":1560221260569},{"name":"C:\\KN\\LP_DEAB\\.browserslistrc","includedInParent":true,"mtime":1559628543032}],"generated":{"js":"module.exports = \"/0f2dec1a1533e2f3c495ff426b84963b.png\";"},"sourceMaps":null,"error":null,"hash":"e8ff5f628a0340a46b39df2bba550eef","cacheData":{"env":{}}}
+{"id":"img/works-mountains-mobile.png","dependencies":[{"name":"C:\\KN\\LP_DEAB\\package.json","includedInParent":true,"mtime":1560387452553},{"name":"C:\\KN\\LP_DEAB\\.browserslistrc","includedInParent":true,"mtime":1560387421364}],"generated":{"js":"module.exports = \"/0f2dec1a1533e2f3c495ff426b84963b.png\";"},"sourceMaps":null,"error":null,"hash":"e8ff5f628a0340a46b39df2bba550eef","cacheData":{"env":{}}}
