+++ src/index.html
@@ -427,7 +427,6 @@
 
   </div>
 
-  <!-- <script src="js/modal.js"></script> -->
   <script src="js/main.js"></script>
 
 </body>
