+++ src/js/modal.js
@@ -85,45 +85,4 @@ document.addEventListener("DOMContentLoaded", function () {
     });
   });
 
-
-  // ------------- Отображать скрытый текст в секции about  -----------------
-
-  // функция отрабатывающая открытие текста и сокрытие кнопки
-  function openText(textSelector, buttonSelector) {
-
-    const moreText = document.getElementById(textSelector);
-    const btnText = document.getElementById(buttonSelector);
-
-    console.log('btn "See more..." was pressed');
-
-    btnText.classList.add('_hide-button');
-    moreText.classList.remove('_hide-text');
-    moreText.classList.add('_visible-text');
-  }
-
-  let btns = document.querySelectorAll('.about-item__see-more');
-
-  for (i = 0; i < btns.length; i++) {
-    let numBtn = 'btn-' + (i + 1);
-    let numTxt = 'more-' + (i + 1);
-
-    document.getElementById(numBtn).addEventListener('click', event => {
-      console.log('btn "' + numBtn + '" was pressed');
-      openText(numTxt, numBtn);
-    });
-
-  };
-
-  // нажали на 1-ю кнопку
-  // document.getElementById("btn-01").addEventListener('click', event => {
-  //   openText('more-01', 'btn-01');
-  // });
-  // // нажали на 2-ю кнопку
-  // document.getElementById("btn-02").addEventListener('click', event => {
-  //   openText('more-02', 'btn-02');
-  // });
-  // // нажали на 2-ю кнопку
-  // document.getElementById("btn-03").addEventListener('click', event => {
-  //   openText('more-03', 'btn-03');
-  // });
 });
